+++ java/org/silvercatcher/reforged/weapons/ReforgedItem.java
@@ -35,6 +35,6 @@ public abstract class ReforgedItem extends Item {
 	
 	public abstract void registerRecipes();
 	
-	public abstract int getHitDamage();
+	public abstract float getHitDamage();
 
 }
